fixed all naming scheme problems for menuitems and orderitems, added up to date seeds

+++ db/migrate/20221011091430_create_orders.rb
@@ -1,7 +1,7 @@
 class CreateOrders < ActiveRecord::Migration[5.2]
   def change
     create_table :orders do |t|
-      t.integer :totalPrice
+      t.decimal :totalPrice
 
       t.timestamps
     end
